requirements.txt
pytest~=6.2.5
requests~=2.32.3
PyYAML~=6.0.2
self~=3.141.0
self~=3.141.0
selenium~=3.141.0